move InterpolateSpectrum to dsp_interpolation.py
