+++ pyfar/dsp/dsp.py
@@ -1,9 +1,7 @@
 import multiprocessing
 import warnings
 import numpy as np
-from scipy.interpolate import interp1d
 from scipy import signal as sgn
-import matplotlib.pyplot as plt
 import pyfar
 from pyfar.dsp import fft
 
@@ -820,283 +818,6 @@ def regularized_spectrum_inversion(
     inverse.freq = np.conj(data) / (np.conj(data)*data + regu_final)
 
     return inverse
-
-
-class InterpolateSpectrum():
-    """
-    Interpolate an incomplete spectrum to a complete single sided spectrum.
-
-    This is intended to interpolate transfer functions, for example sparse
-    spectra that are defined only at octave frequencies or incomplete spectra
-    from numerical simulations.
-
-    Parameters
-    ----------
-    data : FrequencyData
-        Input data to be interpolated. `data.fft_norm` must be `'none'`.
-    method : string
-        Specifies the input data for the interpolation
-
-        ``'complex'``
-            Separate interpolation of the real and imaginary part
-        ``'magnitude_phase'``
-            Separate interpolation if the magnitude and unwrapped phase values
-        ``'magnitude'``
-            Interpolate the magnitude values only. Results in a zero phase
-            signal, which is symmetric around the first sample. This phase
-            response might not be ideal for many applications. Minimum and
-            linear phase responses can be generated with
-            :py:func:`~pyfar.dsp.minimum_phase` and
-            :py:func:`~pyfar.dsp.linear_phase`.
-
-    kind : tuple
-        Three element tuple ``('first', 'second', 'third')`` that specifies the
-        kind of inter/extrapolation below the lowest frequency (first), between
-        the lowest and highest frequency (second), and above the highest
-        frequency (third).
-
-        The individual strings have to be
-
-        ``'zero'``, ``slinear``, ``'quadratic'``, ``'cubic'``
-            Spline interpolation of zeroth, first, second or third order
-        ``'previous'``, ``'next'``
-            Simply return the previous or next value of the point
-        ``'nearest-up'``, ``'nearest'``
-            Differ when interpolating half-integers (e.g. 0.5, 1.5) in that
-            ``'nearest-up'`` rounds up and ``'nearest'`` rounds down.
-
-        The interpolation is done using ``scipy.interpolate.interp1d``.
-    fscale : string, optional
-
-        ``'linear'``
-            Interpolate on a linear frequency axis.
-        ``'log'``
-            Interpolate on a logarithmic frequency axis. Note that 0 Hz can
-            not be interpolated on a logarithmic scale because the logarithm
-            of 0 does not exist. Frequencies of 0 Hz are thus replaced by the
-            next highest frequency before interpolation.
-
-        The default is ``'linear'``.
-    clip : bool, tuple
-        The interpolated magnitude response is clipped to the range specified
-        by this two element tuple. E.g., ``clip=(0, 1)`` will assure that no
-        values smaller than 0 and larger than 1 occur in the interpolated
-        magnitude response. The clipping is applied after the interpolation.
-        The default is ``False`` which does not clip the data.
-
-    Returns
-    -------
-    interpolator : :py:class:`InterpolateSpectrum`
-        The interpolator can be called to interpolate the data (see examples
-        below). It returns a :py:class:`~pyfar.classes.audio.Signal` and has
-        the following parameters
-
-        `n_samples` : int
-            Length of the interpolated time signal in samples
-        `sampling_rate`: int
-            Sampling rate of the output signal in Hz
-        `show` : bool, optional
-            Show a plot of the input and output data. The default is ``False``.
-
-    Examples
-    --------
-    Interpolate a magnitude spectrum, add an artificial linear phase and
-    inspect the results.
-    Note that a similar plot can be created by the interpolator object by
-    ``signal = interpolator(64, 44100, show=True)``
-
-    .. plot::
-
-        >>> import pyfar as pf
-        >>> import matplotlib.pyplot as plt
-        >>> import numpy as np
-        >>> # generate data
-        >>> data = pf.FrequencyData([1, 0], [5e3, 20e3])
-        >>> interpolator = pf.dsp.InterpolateSpectrum(
-        ...     data, 'magnitude', ('nearest', 'linear', 'nearest'))
-        >>> # interpolate 64 samples at a sampling rate of 44100
-        >>> signal = interpolator(64, 44100)
-        >>> # add linear phase
-        >>> signal = pf.dsp.linear_phase(signal, 32)
-        >>> # plot input and output data
-        >>> with pf.plot.context():
-        >>>     _, ax = plt.subplots(2, 2)
-        >>>     # time signal (linear and logarithmic amplitude)
-        >>>     pf.plot.time(signal, ax=ax[0, 0])
-        >>>     pf.plot.time(signal, ax=ax[1, 0], dB=True)
-        >>>     # frequency plot (linear x-axis)
-        >>>     pf.plot.freq(signal, dB=False, freq_scale="linear",
-        ...                  ax=ax[0, 1])
-        >>>     pf.plot.freq(data, dB=False, freq_scale="linear",
-        ...                  ax=ax[0, 1], c='r', ls='', marker='.')
-        >>>     ax[0, 1].set_xlim(0, signal.sampling_rate/2)
-        >>>     # frequency plot (log x-axis)
-        >>>     pf.plot.freq(signal, dB=False, ax=ax[1, 1], label='input')
-        >>>     pf.plot.freq(data, dB=False, ax=ax[1, 1],
-        ...                  c='r', ls='', marker='.', label='output')
-        >>>     min_freq = np.min([signal.sampling_rate / signal.n_samples,
-        ...                        data.frequencies[0]])
-        >>>     ax[1, 1].set_xlim(min_freq, signal.sampling_rate/2)
-        >>>     ax[1, 1].legend(loc='best')
-
-    """
-
-    def __init__(self, data, method, kind, fscale='linear',
-                 clip=False, group_delay=None, unit='samples'):
-
-        # check input ---------------------------------------------------------
-        # ... data
-        if not isinstance(data, pyfar.FrequencyData):
-            raise TypeError('data must be a FrequencyData object.')
-        if data.n_bins < 2:
-            raise ValueError("data.n_bins must be at least 2")
-
-        # ... method
-        methods = ['complex', 'magnitude_phase', 'magnitude']
-        if method not in methods:
-            raise ValueError((f"method is '{method}'' but must be on of the "
-                              f"following: {', '.join(methods)}"))
-
-        # ... kind
-        if not isinstance(kind, tuple) or len(kind) != 3:
-            raise ValueError("kind must be a tuple of length 3")
-        kinds = ['linear', 'nearest', 'nearest-up', 'zero', 'slinear',
-                 'quadratic', 'cubic', 'previous', 'next']
-        for k in kind:
-            if k not in kinds:
-                raise ValueError((f"kind contains '{k}' but must only contain "
-                                  f"the following: {', '.join(kinds)}"))
-
-        # ... fscale
-        if fscale not in ["linear", "log"]:
-            raise ValueError(
-                f"fscale is '{fscale}'' but must be linear or log")
-
-        # ... clip
-        if clip:
-            if not isinstance(clip, tuple) or len(clip) != 2:
-                raise ValueError("clip must be a tuple of length 2")
-
-        # initialize the interpolators ----------------------------------------
-        # store required parameters
-        self._method = method
-        self._clip = clip
-        self._fscale = fscale
-
-        # flatten input data to work with scipy interpolators
-        self._cshape = data.cshape
-        data = data.flatten()
-        self._input = data
-
-        # get the required data for interpolation
-        if method == 'complex':
-            self._data = [np.real(data.freq), np.imag(data.freq)]
-        elif method == 'magnitude_phase':
-            self._data = [np.abs(data.freq),
-                          pyfar.dsp.phase(data, unwrap=True)]
-        else:
-            self._data = [np.abs(data.freq)]
-
-        # frequencies for interpolation (store for testing)
-        self._f_in = self._get_frequencies(data.frequencies.copy())
-
-        # frequency range
-        self._freq_range = [self._f_in[0], self._f_in[-1]]
-
-        # get the interpolators
-        self._interpolators = []
-        for d in self._data:
-            interpolators = []
-            for idx, k in enumerate(kind):
-                if idx == 1:
-                    interpolators.append(interp1d(self._f_in, d, k))
-                else:
-                    interpolators.append(interp1d(
-                        self._f_in, d, k, fill_value="extrapolate"))
-            self._interpolators.append(interpolators)
-
-    def __call__(self, n_samples, sampling_rate, show=False):
-        """
-        Interpolate a Signal with n_samples length.
-        (see class docstring) for more information.
-        """
-
-        # get the query frequencies (store for testing)
-        self._f_query = self._get_frequencies(
-            pyfar.dsp.fft.rfftfreq(n_samples, sampling_rate))
-
-        # get interpolation ranges
-        id_below = self._f_query < self._freq_range[0]
-        id_within = np.logical_and(self._f_query >= self._freq_range[0],
-                                   self._f_query <= self._freq_range[1])
-        id_above = self._f_query > self._freq_range[1]
-
-        # interpolate the data
-        interpolated = []
-        for data in self._interpolators:
-            data_interpolated = np.concatenate((
-                (data[0](self._f_query[id_below])),
-                (data[1](self._f_query[id_within])),
-                (data[2](self._f_query[id_above]))),
-                axis=-1)
-            interpolated.append(data_interpolated)
-
-        # get half sided spectrum
-        if self._method == "complex":
-            freq = interpolated[0] + 1j * interpolated[1]
-        elif self._method == 'magnitude_phase':
-            freq = interpolated[0] * np.exp(-1j * interpolated[1])
-        else:
-            freq = interpolated[0]
-
-        # get initial signal
-        signal = pyfar.Signal(freq, sampling_rate, n_samples, "freq")
-
-        # clip the magnitude
-        if self._clip:
-            signal.freq = np.clip(
-                np.abs(signal.freq),
-                self._clip[0],
-                self._clip[1]) * np.exp(-1j * phase(signal))
-
-        if show:
-            # plot input and output data
-            with pyfar.plot.context():
-                _, ax = plt.subplots(2, 2)
-                # time signal (linear amplitude)
-                pyfar.plot.time(signal, ax=ax[0, 0])
-                # time signal (log amplitude)
-                pyfar.plot.time(signal, ax=ax[1, 0], dB=True)
-                # frequency plot (linear x-axis)
-                pyfar.plot.freq(signal, dB=False, freq_scale="linear",
-                                ax=ax[0, 1])
-                pyfar.plot.freq(self._input, dB=False, freq_scale="linear",
-                                ax=ax[0, 1], c='r', ls='', marker='.')
-                ax[0, 1].set_xlim(0, sampling_rate/2)
-                # frequency plot (log x-axis)
-                pyfar.plot.freq(signal, dB=False, ax=ax[1, 1], label='input')
-                pyfar.plot.freq(self._input, dB=False, ax=ax[1, 1],
-                                c='r', ls='', marker='.', label='output')
-                min_freq = np.min([sampling_rate / n_samples,
-                                   self._input.frequencies[0]])
-                ax[1, 1].set_xlim(min_freq, sampling_rate/2)
-                ax[1, 1].legend(loc='best')
-
-        return signal
-
-    def _get_frequencies(self, frequencies):
-        """
-        Return frequencies for creating or quering interpolation objects.
-
-        In case logfrequencies are requested, 0 Hz entries are replaced by
-        the next highest frequency, because the logarithm of 0 does not exist.
-        """
-        if self._fscale == "log":
-            if frequencies[0] == 0:
-                frequencies[0] = frequencies[1]
-            frequencies = np.log(frequencies)
-
-        return frequencies
 
 
 def _cross_fade(first, second, indices):
